Adição de Assets na cena de Revisão do Gameplay. Balanceamento de range, velocidade de personagens e torre. Remoção de arvores e respawn points de gemas. Limpeza de código.

+++ Assets/MMagicRush/Scripts/BaseDefense.cs
@@ -10,7 +10,7 @@ public class BaseDefense : MonoBehaviour {
 
 	public int team;
 
-	public int reach;
+	public float reach;
 	public float shootCD;
 	public float damageSpeed;
 	public float danoCD;
